Reduce the magnitude of the winDescent value (#271)
This PR modifies the winDescent value to align with the sTypo and hhea descender values. 

## PR Checklist
* [x] Closes #269 
* [x] CLA signed.

## Detailed Description of the Pull Request / Additional comments
Due to the way that Notepad++ determines the bounding box of the CR and LF special characters, the background black box was not rendering correctly. Modifying the winDescent value resolves this issue and does not appear to cause any other problems.

## Validation Steps Performed
- Evaluated in Notepad++ and other text editors. 
- Determined that the only glyphs which extend past the sTypo descender are the box drawing characters, so clipping is not problematic.