Make Cascadia Code a variable font w/ 6 named variations (#304)
This major update adds a weight range to Cascadia Code. This font is now
being built as a Variable Font, which enables users to select the
perfect weight for their preference. 

The weight range now extends from ExtraLight (200) to Bold (700), with
the current Regular set at 500 (400 in the font selector, 500
internally). As a variable font, OS / rendering engine support may vary.
Users running Windows 10 and Windows Terminal will have access to the
full range of font weights. Other applications may only have access to
the named instances (ExtraLight / Semilight / Light / Regular / SemiBold
/ Bold) depending on inbuilt support. 

Static instance OTFs are also provided. At current, static TTFs are not
built in this update, but it is something we will consider in the
future. 

Variation implementation tested on Windows and Mac, font hinted and
reviewed on high and low DPI devices. 

Closes #25 - weight axis added
Closes #43 - with weight addition, parenthesis width is preserved 
Closes #284 - ligature now broken for easier recognition
Closes #90 - produced as a variable font
Closes #128 - MORE WEIGHTS 
Closes #285 - contextual code removed

Signed-off-by: Aaron Bell <aaron@sajatypeworks.com>