Update Bold and other smaller fixes (#311)
This PR shifts the position of the "Bold" instance closer to the Regular
in weight.

Additionally, I've fixed a couple of other smaller issues around
diacritic marks and symbols.

In conducting research into Visual Studio's rendering of Cascadia Code
(versus other fonts), I began to feel that the Bold that was in use felt
a bit too heavy. In selecting a lighter point on the axis, I was able to
achieve better rendering results in VS and without drawing as much
attention to the Bold. 

The main issue with VS is that it does not support y-direction
smoothing, so overly bold Glyphs start to look quite distorted and
weird. Additionally, it offers only minimal variable font support so I
can’t STAT my way out of it ;).

Other fixes:
* Added linked bolds for the lighter weights.
* Tweaked design of primes to give a more angled feel.
* Added weight to the top of the acutecomb and gravecomb glyphs
* Added anchors to the common lowercase Greek characters. 
* Corrected build script to correctly build TTF variable fonts. 
* Resolved some introduced hinting bugs

VALIDATION

Reviewed in VS, VS Code, notepad. Validated no impact on hinting
instructions. 

Closes #294 - changed design of primes to make them appear more angled.
Closes #282 - added anchors to Greek for use in Julia REPL
Closes #268 - added additional weighting to grave / acute to help
differentiate from dot