Add OVERLAP flag to simple/compound glyphs, fix interpl. error (#352)
This commit modifies the build script to add "OVERLAP_SIMPLE" to normal
glyphs and "OVERLAP_COMPOUND" to component glyphs. 

Under FreeType, "overlapping contours can produce pixels with inflated
coverage because FreeType normally integrates the coverage. To mitigate
the artefacts, oversampling is necessary (increase resolution 4x4 and
average the results)" ([mailinglist quote]). FreeType relies on the
"OVERLAP_SIMPLE" and "OVERLAP_COMPOUND" flags to be set to know when
this oversampling should take place. 

At current, Cascadia does not contain either of these two flags. This
fix will add one or the other to every glyph in Cascadia (depending on
type of glyph), thus resolving the issue. Once the ufo2ft library adds
functionality to address this issue, the build script modification can
be removed. 

The addition of these flags has no impact on the behavior of the font in
Mac or Windows environments. 

Closes #350
Closes #355 - switches contour order to resolve interpolation error at wght 300

[mailinglist quote]: https://savannah.nongnu.org/bugs/index.php?59026