Correct mark positioning for Salishan (#364)
Properly adds the new marks required for Salishan to the GDEF feature
for marks. (Thanks to Timothy Montler for catching this) 

The special feature files that override as part of the font building
process weren't updated properly to include the new diacritical marks,
so they were not being aligned properly. This has now been corrected.