Set the names of the static instances properly (#370)
As far as I can tell, what essentially happened here is that _instances_
in the designspace were not being renamed, even though the sources were
being renamed. As a result, the variable fonts would build correctly,
but the static instances were not. The font naming approach has now been
changed and is producing reliable results. 

What's particularly vexing is that for the life of me I can't figure out
how the code actually worked before, because as far as I can tell, we've
_never_ modified the instance code before.

Fixes #369.

Co-authored-by: Aaron Bell <aaron@sajatypeworks.com>