Fixing images (#516)
Fixing the image samples on the main page (it appears it was using an older version of the font). Also found a couple of bugs in the source for the graphical squares (wherein the components are mirrored). This has no impact on font output but is better for the source.