Fix NF Mono glyph selection, NF variants STAT, PL rclt table (#721)
The build issue I was running into was caused by the logic that chooses
when to merge Powerline glyphs into the final set before building.

It covered the following conditions:

| Conditions | Outcome                         |
| ---------- | ------------------------------- |
| Mono       | Set name only ✅                |
| Mono + PL  | Set name, merge PL ✅           |
| PL         | Set name, merge PL ✅           |
| NF         | Set name, merge PL, merge NF ✅ |

However, it missed one:

| Conditions | Outcome                         |
| ---------- | ------------------------------- |
| Mono + NF  | ❔❔❔                          | 

I've simplified the logic so that we don't make this mistake again. All
non-Cascadia-Code variants will have their names overridden, and if they
contain PL or NF they will get special treatment.

I've also added the NF variants to the STAT table table and fixed the
generation of rclt for PL variants.