Complete the branch rename from master to main (again) (#734)
Non-italic license strings were updated in Oct. 2020, but italic
versions were not.
This commit updates the license string of the remaining italic versions,
completing the branch rename from master to main (for good this time).

## Detailed Description of the Pull Request / Additional comments

I only noticed this when cleaning up custom builds and installing the
2404.23 release, and now I can't unsee it.
Since I still had a fork, I figured I could submit a PR just as fast as
submitting an issue. So here it is.

During the October 2020 branch rename from master to main, the license
strings were updated in non-italic fonts, but not in italic ones.

This PR updates the license string in the italic fonts to match the
non-italic ones.