Update all packages to their latest versions (#726)
I've run builds with this configuration plus Python 3.11; it appears to
work.

In part, this is required for our build agents - they use a version of
macOS that no longer runs Python 3.9, or binary packages have aged out,
or something. It's not ideal. But... we can check the fonts (I suppose)
after we build them.